theshow/s1/e4/requirements.txt
Pillow
python-gnupg
